backend/requirements.txt
fastapi
uvicorn
requests
google-generativeai
langchain
langchain-google-genai
python-dotenv
